requirements.txt
torch
torchvision
tensorboard
scipy==1.13.1
numpy==1.24.3
matplotlib==3.9.4
opencv_python==4.10.0.84
tqdm==4.67.1
Pillow==9.4.0
h5py==2.10.0
